LPS-127488 Add missing propsTransformer for <clay:button />
In [LPS-125256](https://issues.liferay.com/browse/LPS-125256), we forgot to add
a `propsTransformer` for the button, which caused a loss of functionality.

With this change, it's now possible to choose the layout for navigation menus, as expected.

**Test plan**:
- Navigate to **Site Builder**>**Pages**
- Create a new public page based on the Blank template (found in **Basic
Tempates**)
- Publish the page
- Navigate to **Site Builder**>**Navigation Menus**
- Create a new navigation menu
- From the "New" button select "Page"
- In the item selector, select the page you just published
- Select the page by clicking on it
- Change the layout by clicking on the "Choose" button
